+++ config/src/config.rs
@@ -3,24 +3,21 @@
 
 use crate::{
     keys::{ConsensusKeyPair, NetworkKeyPairs},
-    seed_peers::{SeedPeersConfig, SeedPeersConfigHelpers},
-    trusted_peers::{ConfigHelpers, ConsensusPrivateKey, NetworkPeersConfig, NetworkPrivateKeys},
-    utils::{deserialize_whitelist, get_available_port, get_local_ip, serialize_whitelist},
+    seed_peers::SeedPeersConfigHelpers,
+    trusted_peers::{ConfigHelpers, ConsensusPrivateKey, NetworkPrivateKeys},
+    utils::{deserialize_whitelist, get_available_port, serialize_whitelist},
 };
 use failure::prelude::*;
-use libra_crypto::ValidKey;
 use libra_tools::tempdir::TempPath;
 use libra_types::{
     transaction::{SignedTransaction, Transaction, SCRIPT_HASH_LENGTH},
     PeerId,
 };
-use parity_multiaddr::Multiaddr;
 use prost::Message;
 use serde::{de::DeserializeOwned, Deserialize, Serialize};
 use std::{
     collections::HashSet,
     convert::TryFrom,
-    fmt,
     fs::File,
     io::{Read, Write},
     path::{Path, PathBuf},
@@ -39,6 +36,8 @@ mod execution_config;
 pub use execution_config::*;
 mod metrics_config;
 pub use metrics_config::*;
+mod network_config;
+pub use network_config::*;
 mod state_sync_config;
 pub use state_sync_config::*;
 mod storage_config;
@@ -105,34 +104,6 @@ impl Default for BaseConfig {
     }
 }
 
-#[derive(Copy, Clone, Debug, Eq, PartialEq)]
-pub enum RoleType {
-    Validator,
-    FullNode,
-}
-
-impl<T> std::convert::From<T> for RoleType
-where
-    T: AsRef<str>,
-{
-    fn from(t: T) -> RoleType {
-        match t.as_ref() {
-            "validator" => RoleType::Validator,
-            "full_node" => RoleType::FullNode,
-            _ => unimplemented!("Invalid node role: {}", t.as_ref()),
-        }
-    }
-}
-
-impl fmt::Display for RoleType {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        match self {
-            RoleType::Validator => write!(f, "validator"),
-            RoleType::FullNode => write!(f, "full_node"),
-        }
-    }
-}
-
 impl BaseConfig {
     /// Constructs a new BaseConfig with an empty temp directory
     pub fn new(data_dir_path: PathBuf) -> Self {
@@ -168,101 +139,6 @@ impl Default for LoggerConfig {
             is_async: true,
             chan_size: 256,
         }
-    }
-}
-
-#[cfg_attr(any(test, feature = "fuzzing"), derive(Clone))]
-#[derive(Debug, Deserialize, Serialize)]
-#[serde(default)]
-pub struct NetworkConfig {
-    pub peer_id: String,
-    // TODO: Add support for multiple listen/advertised addresses in config.
-    // The address that this node is listening on for new connections.
-    pub listen_address: Multiaddr,
-    // The address that this node advertises to other nodes for the discovery protocol.
-    pub advertised_address: Multiaddr,
-    pub discovery_interval_ms: u64,
-    pub connectivity_check_interval_ms: u64,
-    // Flag to toggle if Noise is used for encryption and authentication.
-    pub enable_encryption_and_authentication: bool,
-    // If the network is permissioned, only trusted peers are allowed to connect. Otherwise, any
-    // node can connect. If this flag is set to true, the `enable_encryption_and_authentication`
-    // must also be set to true.
-    pub is_permissioned: bool,
-    // The role of the node in the network. One of: {"validator", "full_node"}.
-    pub role: String,
-    // network_keypairs contains the node's network keypairs.
-    // it is filled later on from network_keypairs_file.
-    #[serde(skip)]
-    pub network_keypairs: NetworkKeyPairs,
-    pub network_keypairs_file: PathBuf,
-    // network peers are the nodes allowed to connect when the network is started in permissioned
-    // mode.
-    #[serde(skip)]
-    pub network_peers: NetworkPeersConfig,
-    pub network_peers_file: PathBuf,
-    // seed_peers act as seed nodes for the discovery protocol.
-    #[serde(skip)]
-    pub seed_peers: SeedPeersConfig,
-    pub seed_peers_file: PathBuf,
-}
-
-impl Default for NetworkConfig {
-    fn default() -> NetworkConfig {
-        NetworkConfig {
-            peer_id: "".to_string(),
-            role: "validator".to_string(),
-            listen_address: "/ip4/0.0.0.0/tcp/6180".parse::<Multiaddr>().unwrap(),
-            advertised_address: "/ip4/127.0.0.1/tcp/6180".parse::<Multiaddr>().unwrap(),
-            discovery_interval_ms: 1000,
-            connectivity_check_interval_ms: 5000,
-            enable_encryption_and_authentication: true,
-            is_permissioned: true,
-            network_keypairs_file: PathBuf::from("network_keypairs.config.toml"),
-            network_keypairs: NetworkKeyPairs::default(),
-            network_peers_file: PathBuf::from("network_peers.config.toml"),
-            network_peers: NetworkPeersConfig::default(),
-            seed_peers_file: PathBuf::from("seed_peers.config.toml"),
-            seed_peers: SeedPeersConfig::default(),
-        }
-    }
-}
-
-impl NetworkConfig {
-    pub fn load<P: AsRef<Path>>(&mut self, path: P) -> Result<()> {
-        if !self.network_peers_file.as_os_str().is_empty() {
-            self.network_peers = NetworkPeersConfig::load_config(
-                path.as_ref().with_file_name(&self.network_peers_file),
-            );
-        }
-        if !self.network_keypairs_file.as_os_str().is_empty() {
-            self.network_keypairs = NetworkKeyPairs::load_config(
-                path.as_ref().with_file_name(&self.network_keypairs_file),
-            );
-        }
-        if !self.seed_peers_file.as_os_str().is_empty() {
-            self.seed_peers =
-                SeedPeersConfig::load_config(path.as_ref().with_file_name(&self.seed_peers_file));
-        }
-        if self.advertised_address.to_string().is_empty() {
-            self.advertised_address =
-                get_local_ip().ok_or_else(|| ::failure::err_msg("No local IP"))?;
-        }
-        if self.listen_address.to_string().is_empty() {
-            self.listen_address =
-                get_local_ip().ok_or_else(|| ::failure::err_msg("No local IP"))?;
-        }
-        // If PeerId is not set, it is derived from NetworkIdentityKey.
-        if self.peer_id == "" {
-            self.peer_id = PeerId::try_from(
-                self.network_keypairs
-                    .get_network_identity_public()
-                    .to_bytes(),
-            )
-            .unwrap()
-            .to_string();
-        }
-        Ok(())
     }
 }
